# 8255949: AArch64: Add support for vectorized shift right and accumulate