@dgbo 2020-11-10
@theRealAph Thanks for the review.
/integrate

I'll fix the register naming style of Base64.encode intrinisc in that PR as suggested.